create_unix should create an unlinked File
Allows the OS to clean up the file on disk after the last file
descriptor is closed. This used to be the behaviour of create_unix but
this was lost after porting to use the util helper in
 bde363d751db4247296ae9345c2e346ec3abe971.

Does not affect linux on systems with O_TMPFILE available.

Add test to assert that temporary files are indeed removed after use.