fix: don't claim 'too many temporary files exist' if a file already existed, once.
Previously when creating a temporary file without any random characters
in it, a loop would effectively run once and fail with the conclusion
that a file could not be created because too many temporary files exist
(2_147_483_648).

Now it will return the original failure, which typically is that the
file with the given name already existed.

This can more easily happen on case-insentiive file systems, and the
correct error message helps understanding what went wrong.

See https://github.com/Byron/gitoxide/issues/595 for reference.