NamedTempFile: Fix infinite recursion in trait bounds (#225)
Due to bug rust-lang/rust#96634, these generic bounds may cause an
infinite recursion in the compiler, even in unrelated code.

This specializes the `Write for &NamedTempFile<F>` impl on `File`
instead. This keeps the API the same as before the generic parameter `F`
was added, so it shouldn't break any code.