Provide implementations for I/O safety traits unconditionally (#244)
* Provide implementations for I/O safety traits unconditionally

These traits have been stabilized with Rust 1.63. Since that is now our MSRV, we can drop the
`build.rs` as well as the `autocfg` build dependency.
While at it, I've also added a few `use` statements to make the implemenations more readable.

* Make `AsFd`/`AsRawFd` implementations available on wasi

`std` provides these traits on wasi, so we should probably do so as well.