fix(windows): don't automatically delete files on open
We tried doing this to match the behavior on Unix platforms but this
appears to trigger some form of Windows kernel bug at scale. Given that
it's not necessary, we're reverting it. I'm leaving the feature flag in
the `Cargo.toml` for now but will remove it in the near future.

fixes #339