feat: make it possible to disable cleanup in-place after creation (#355)
This makes it possible to disable automatic cleanup of temporary files &
directories _after_ they've been created.

1. Rename `Builder::keep` into `Builder::disable_cleanup` for
consistency. `::keep` on `NamedTempDir`, `TempDir`, `TempPath`, etc.
provide a way to _convert_ (one-way) a temporary directory/file into a
persistent temporary directory/file but aren't a way to simply disable
automatic cleanup in-place.
2. Add `disable_cleanup` methods to all named temporary file types,
making it possible to replicate the behavior of
`Builder::disable_cleanup` _after_ creating a temporary file. Given that
this operation is in-place, it can be undone.

fixes #347